+++ Tests/Units/Collection/HydraPaginatedCollection.php
@@ -5,7 +5,7 @@ namespace Mapado\RestClientSdk\Tests\Units\Collection;
 use atoum;
 
 /**
- * SdkClient
+ * HydraPaginatedCollection
  *
  * @uses   atoum
  * @author Florent Clerc <florent.clerc@mapado.com>
@@ -24,7 +24,7 @@ class HydraPaginatedCollection extends atoum
         $json = json_decode(file_get_contents(__DIR__ . '/../../data/ticketing.list.paginated.json'), true);
 
         $this
-            ->given($collection = new \Mapado\RestClientSdk\Collection\HydraPaginatedCollection($json))
+            ->given($collection = new \Mapado\RestClientSdk\Collection\HydraPaginatedCollection($json['hydra:member'], $json))
 
             ->then
             ->object($collection)